requirements.txt
ipython
pandas
requests
streamlit
pytest-playwright
watchdog
matplotlib
numpy
click
psycopg2-binary
SQLAlchemy
